% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{emv}
\alias{emv}
\title{EMV}
\usage{
emv(mat, prop)
}
\arguments{
\item{mat}{this is alternative and state of State of Nature payoff table}

\item{prop}{this is probability vecotr for each state of nature}
}
\value{
emv
}
\description{
The Expected Monetary Value (EMV) of a decision is the
long-run average value of the outcome of that decision.
In other words, if we have a decision to make, let's
suppose taht we could make that exact same decicion
under the exact same circumstances many, many times
(obviously, we can't in real life, but suppose we could).
One time a good State of nature may occur, and we would
have a very positive outcome. Another time we many have a
negative outcome becuase some less-favorable State of
Nature happened. If, somehow, we could repeat that
decision lots and lots of times, and determine the
outcome for each time, and then average all those
outcomes, then we would have the EMV of the decision alternative.
}
