% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{EOL}
\alias{EOL}
\title{EOL}
\usage{
EOL(mat, prop)
}
\arguments{
\item{mat}{this is alternative and state of State of Nature payoff table}

\item{prop}{this is probability vecotr for each state of nature}
}
\value{
EOL value vector
}
\description{
This calculates the expected opportunity loss
}
