% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{EVPI}
\alias{EVPI}
\title{EVPI}
\usage{
EVPI(mat, prop)
}
\arguments{
\item{mat}{this is alternative and state of State of Nature payoff table}

\item{prop}{this is probability vecotr for each state of nature}
}
\value{
EVPI
}
\description{
This calculates the expected value of perfect information.
}
