% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{EVwPI}
\alias{EVwPI}
\title{EVwPI}
\usage{
EVwPI(mat, prop)
}
\arguments{
\item{mat}{this is alternative and state of State of Nature payoff table}

\item{prop}{this is probability vecotr for each state of nature}
}
\value{
EV with PI EV|PI
}
\description{
This calculates the expected value with perfect information.
}
