% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{hurwicz}
\alias{hurwicz}
\title{Hurwicz}
\usage{
hurwicz(mat, alpha)
}
\arguments{
\item{mat}{this is alternative and state of State of Nature payoff table}

\item{alpha}{the Hurwicz Alpha scaler between 0 and 1.}
}
\value{
EVPI
}
\description{
The Hurwicz criterion is arguably one of the most widely used rules
in decision-making under uncertainty. It allows the decision maker
to simultaneously take into account the best and the worst possible
outcomes, by articulating a "coefficient of optimism" that determines
the emphasis on the best end.
}
