% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{maximax}
\alias{maximax}
\title{maximax}
\usage{
maximax(mat)
}
\arguments{
\item{mat}{this is alternative and state of State of Nature payoff table}
}
\value{
maximax value (not the location)
}
\description{
This function calculated the maximax criterion
Maximax criterion – this decision criterion is
applied by the most optimist decision make
when he thinks optimistically about the
happening of events affecting decision.
If this philosophy is followed, the manager
will chose that alternative under which it
is possible to receive the most favourable
pay-off.
}
