% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{maximin}
\alias{maximin}
\title{maximin}
\usage{
maximin(mat)
}
\arguments{
\item{mat}{this is alternative and state of State of Nature 1 payoff table}
}
\value{
maximin value (not the location)
}
\description{
Maximin criterion- this criterion is adopted by the
most pessimistic decision maker. The manager believes
that worst possible may take place.  This pessimism
cause the selection of that alternative which
maximizes the least favourable  payoff.
}
