% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{minEOL}
\alias{minEOL}
\title{minEOL}
\usage{
minEOL(mat, prop)
}
\arguments{
\item{mat}{this is alternative and state of State of Nature payoff table}

\item{prop}{this is probability vecotr for each state of nature}
}
\value{
minimum of the EOL vector
}
\description{
This calculates the minimum expected opportunity loss
}
