% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{minimax}
\alias{minimax}
\title{minimax}
\usage{
minimax(mat)
}
\arguments{
\item{mat}{this is alternative and state of State of Nature 1 payoff table}
}
\value{
minimax value (not the location)
}
\description{
Minimax criterion- minimax criterion leads to the
minimization of regret. The managerial regret is
define as the pay off  for each alternative under
state of nature of compitative action subtracted from
the most favourable payoff which is possible with the
happening of the particular event. When manager chooses
an alternatives and when a state of nature takes place
which does not result in the most favourable payoff,
regret takes place.
}
