% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{regretMatrix}
\alias{regretMatrix}
\title{regretMatrix}
\usage{
regretMatrix(mat)
}
\arguments{
\item{mat}{this is alternative and state of State of Nature payoff table}
}
\value{
regret Matrix (matrix object)
}
\description{
This calculates the regret matrix from the payoff table.
}
