% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{regretminimax}
\alias{regretminimax}
\title{regretminimax}
\usage{
regretminimax(mat)
}
\arguments{
\item{mat}{this is alternative and state of State of Nature payoff table}
}
\value{
minimax of the EOL
}
\description{
This calculates the minimax of the regret matrix
}
