% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{treeRollup}
\alias{treeRollup}
\title{treeRollUp}
\usage{
treeRollup(mat, prop)
}
\arguments{
\item{mat}{this is alternative and state of State of Nature payoff table}

\item{prop}{this is probability vecotr for each state of nature}
}
\value{
probability
}
\description{
This calculates the probability of a tree roll up, or (as I call it)
a straight probability calculation
}
